feat(all): add support for module name suggestion
module name suggestion was first implemented as a macro for templates,
then I changed my mind to implement it as a CompletionContributor which adds LookupElements
for module names in `angular.module("moduleName",...)`
next step would be to suggest module names also in dependencies array in module registration.
it's worth mentioning that module name suggestion as a macro is also still available under macro name
`suggestAngularModule`.
For module suggestion to be used in both macro and completionContributor, the logic of suggestion is
defined and implemented in AngularModuleNameSuggester interface and AngularModuleNameSuggesterImpl
respectively.